Add support for codegen of "unknown" type names (#273)
* Add support for codegen of "unknown" type names

If the analyzer is unable to resolve a type, it reports that the type is
"dynamic", and our codegen writes it out way. This situation is
particularly noticeable because the build package's analyzer draws from
a pre-computed analysis summary that does not include dart:ui.

The recent work on annotated stores demonstrates a similar and related
problem, where we'd like to reference generated types in our store
definitions, but because the analyzer doesn't know about types that have
not been generated, dynamic litters the generated code.

This PR explores an approach where we output types to generated code
as they were written in the input, instead of relying on the analyzer's
interpretation. If we encounter a type that the analyzer reports as
dynamic, we look up the element's AST node, and use the name found
there. Magic.

* Improve test coverage

* Increase code coverage