Use Element model to supply type names
Now that dart:ui types are natively supported by the `build` library,
we can move back to using the Element model to supply type names.

Of note, the analyzer's type inferrence capabilities will now be
available for use by our code generator.

Example:
```dart
class StoreBase implements Store {
  @observable
  var someField = '12345';
}
```

In this example, `someField` will be recognized as a `String`, despite
the lack of a type annotation.

This PR also removes the dart:ui test, as handling them is no longer a
concern of this library.
[`build_resolvers`](https://pub.dev/packages/build_resolvers#120)
handles it appropriately when the generator runs in a Flutter project.

Fixes #324

MR: element_type_names