Perf/increase performance when spy is disabled (#723)
* perf: avoid calculating time on disabled spy

When `spy` is disabled we should not use `DateTime.now()` as it has been
found to be a very expensive operation. In this case, the duration of
the event defaults to zero.

* perf: avoid creating strings on disabled spy

We should not create Strings which we're not going to use. In the
counter example we never get to the `conditionallyRunInAction`'s else,
therefore never using the memory and putting more pressure on the GC.

This commit reduces the usage of memory by not creating `actionName`
if we have spy disabled.

* fix: mocks isSpyEnabled on tests

* fix: uncomment expectations in tests

* fix: minor linting issue in master

* Update mobx/lib/src/core/atom_extensions.dart

Co-authored-by: Ascênio <Ascenio@users.noreply.github.com>

Co-authored-by: Pavan Podila <pavanpodila@users.noreply.github.com>